Rename Flatten -> flatten
