+++ learning/katas/go/core_transforms/flatten/flatten/pkg/task/task.go
