+++ learning/katas/go/core_transforms/flatten/flatten/task-remote-info.yaml
