+++ learning/katas/go/core_transforms/flatten/lesson-info.yaml
@@ -18,4 +18,4 @@
 #
 
 content:
-- Flatten
+- flatten
