+++ learning/katas/go/core_transforms/flatten/lesson-remote-info.yaml
